Add initial Provision interface and SpireHelm implementation
Adds a Provision interface that abstracts the provisioning of workload
identity clusters.

Refactors the cofidectl up and down implementations into a SpireHelm
type that implements the Provision interface.

This is the first stage of refactoring for #56 and does not change the
implementation. Notably, the UI (status spinner) remains in the
SpireHelm implementation, but will be extracted in a future change.
