+++ cmd/cofidectl/cmd/down.go
@@ -4,13 +4,7 @@
 package cmd
 
 import (
-	"context"
-	"fmt"
-
-	trust_zone_proto "github.com/cofide/cofide-api-sdk/gen/go/proto/trust_zone/v1alpha1"
 	cmdcontext "github.com/cofide/cofidectl/pkg/cmd/context"
-	"github.com/cofide/cofidectl/cmd/cofidectl/cmd/statusspinner"
-	"github.com/cofide/cofidectl/pkg/provider/helm"
 	"github.com/spf13/cobra"
 )
 
@@ -39,38 +33,9 @@ func (d *DownCommand) DownCmd() *cobra.Command {
 				return err
 			}
 
-			trustZones, err := ds.ListTrustZones()
-			if err != nil {
-				return err
-			}
-
-			if len(trustZones) == 0 {
-				fmt.Println("no trust zones have been configured")
-				return nil
-			}
-
-			if err := uninstallSPIREStack(cmd.Context(), trustZones); err != nil {
-				return err
-			}
-
-			return nil
+			provision := d.cmdCtx.PluginManager.GetProvision()
+			return provision.TearDown(cmd.Context(), ds)
 		},
 	}
 	return cmd
-}
-
-func uninstallSPIREStack(ctx context.Context, trustZones []*trust_zone_proto.TrustZone) error {
-	for _, trustZone := range trustZones {
-		prov, err := helm.NewHelmSPIREProvider(ctx, trustZone, nil, nil)
-		if err != nil {
-			return err
-		}
-
-		s := statusspinner.New()
-		statusCh := prov.ExecuteUninstall()
-		if err := s.Watch(statusCh); err != nil {
-			return fmt.Errorf("uninstallation failed: %w", err)
-		}
-	}
-	return nil
 }
